requirements.lock.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
arrow==1.2.3
attrs==25.3.0
black==25.9.0
blinker==1.9.0
boto3==1.40.50
botocore==1.40.50
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.1.3
cssselect2==0.8.0
decorator==5.2.1
Deprecated==1.2.18
distlib==0.4.0
distro==1.9.0
fastapi==0.116.1
ffmpeg-python==0.2.0
filelock==3.19.1
flake8==7.3.0
fonttools==4.59.1
frozenlist==1.8.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-analytics-data==0.18.19
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.43.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
identify==2.6.14
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
isort==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
line-bot-sdk==3.19.1
Markdown==3.8.2
MarkupSafe==3.0.2
mccabe==0.7.0
moviepy==2.2.1
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.5.0
nodeenv==1.9.1
numpy==2.3.3
oauthlib==3.3.1
openai==2.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
proglog==0.1.12
propcache==0.4.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
pyflakes==3.4.0
Pygments==2.19.2
pypandoc==1.15
pyparsing==3.2.3
pyphen==0.17.2
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
selenium==4.36.0
sh==1.14.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spark-parser==1.9.0
starlette==0.47.3
streamlit==1.50.0
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uncompyle6==3.9.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
watchdog==6.0.0
weasyprint==66.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
xdis==6.1.5
yarl==1.22.0
zopfli==0.2.3.post1
